refactor: moved internal packages from pkg to internal
